+++ lib/getLinesToRender.worker.js
@@ -1,17 +1,7 @@
-import getLinesToRender from "./getLinesToRender";
-console.log("Loaded RDV Worker");
-self.onmessage = function (e) {
-    try {
-        var argumentObject = e.data;
-        // Example: simple CPU-heavy simulation
-        var result = getLinesToRender(argumentObject);
-        // Send the result back to the main thread
-        self.postMessage({ success: true, data: result });
-    }
-    catch (error) {
-        self.postMessage({ success: false, error: error.message });
-    }
-    finally {
-        self.close();
-    }
-};
+(()=>{var x=class{diff(e,n,i={}){let r;typeof i=="function"?(r=i,i={}):"callback"in i&&(r=i.callback);let o=this.castInput(e,i),f=this.castInput(n,i),s=this.removeEmpty(this.tokenize(o,i)),l=this.removeEmpty(this.tokenize(f,i));return this.diffWithOptionsObj(s,l,i,r)}diffWithOptionsObj(e,n,i,r){var o;let f=d=>{if(d=this.postProcess(d,i),r){setTimeout(function(){r(d)},0);return}else return d},s=n.length,l=e.length,c=1,u=s+l;i.maxEditLength!=null&&(u=Math.min(u,i.maxEditLength));let a=(o=i.timeout)!==null&&o!==void 0?o:1/0,D=Date.now()+a,h=[{oldPos:-1,lastComponent:void 0}],m=this.extractCommon(h[0],n,e,0,i);if(h[0].oldPos+1>=l&&m+1>=s)return f(this.buildValues(h[0].lastComponent,n,e));let L=-1/0,g=1/0,y=()=>{for(let d=Math.max(L,-c);d<=Math.min(g,c);d+=2){let P,C=h[d-1],w=h[d+1];C&&(h[d-1]=void 0);let v=!1;if(w){let E=w.oldPos-d;v=w&&0<=E&&E<s}let p=C&&C.oldPos+1<l;if(!v&&!p){h[d]=void 0;continue}if(!p||v&&C.oldPos<w.oldPos?P=this.addToPath(w,!0,!1,0,i):P=this.addToPath(C,!1,!0,1,i),m=this.extractCommon(P,n,e,d,i),P.oldPos+1>=l&&m+1>=s)return f(this.buildValues(P.lastComponent,n,e))||!0;h[d]=P,P.oldPos+1>=l&&(g=Math.min(g,d-1)),m+1>=s&&(L=Math.max(L,d+1))}c++};if(r)(function d(){setTimeout(function(){if(c>u||Date.now()>D)return r(void 0);y()||d()},0)})();else for(;c<=u&&Date.now()<=D;){let d=y();if(d)return d}}addToPath(e,n,i,r,o){let f=e.lastComponent;return f&&!o.oneChangePerToken&&f.added===n&&f.removed===i?{oldPos:e.oldPos+r,lastComponent:{count:f.count+1,added:n,removed:i,previousComponent:f.previousComponent}}:{oldPos:e.oldPos+r,lastComponent:{count:1,added:n,removed:i,previousComponent:f}}}extractCommon(e,n,i,r,o){let f=n.length,s=i.length,l=e.oldPos,c=l-r,u=0;for(;c+1<f&&l+1<s&&this.equals(i[l+1],n[c+1],o);)c++,l++,u++,o.oneChangePerToken&&(e.lastComponent={count:1,previousComponent:e.lastComponent,added:!1,removed:!1});return u&&!o.oneChangePerToken&&(e.lastComponent={count:u,previousComponent:e.lastComponent,added:!1,removed:!1}),e.oldPos=l,c}equals(e,n,i){return i.comparator?i.comparator(e,n):e===n||!!i.ignoreCase&&e.toLowerCase()===n.toLowerCase()}removeEmpty(e){let n=[];for(let i=0;i<e.length;i++)e[i]&&n.push(e[i]);return n}castInput(e,n){return e}tokenize(e,n){return Array.from(e)}join(e){return e.join("")}postProcess(e,n){return e}get useLongestToken(){return!1}buildValues(e,n,i){let r=[],o;for(;e;)r.push(e),o=e.previousComponent,delete e.previousComponent,e=o;r.reverse();let f=r.length,s=0,l=0,c=0;for(;s<f;s++){let u=r[s];if(u.removed)u.value=this.join(i.slice(c,c+u.count)),c+=u.count;else{if(!u.added&&this.useLongestToken){let a=n.slice(l,l+u.count);a=a.map(function(D,h){let m=i[c+h];return m.length>D.length?m:D}),u.value=this.join(a)}else u.value=this.join(n.slice(l,l+u.count));l+=u.count,u.added||(c+=u.count)}}return r}};var R=class extends x{},K=new R;function j(t,e,n){return K.diff(t,e,n)}function F(t,e){let n;for(n=0;n<t.length&&n<e.length;n++)if(t[n]!=e[n])return t.slice(0,n);return t.slice(0,n)}function M(t,e){let n;if(!t||!e||t[t.length-1]!=e[e.length-1])return"";for(n=0;n<t.length&&n<e.length;n++)if(t[t.length-(n+1)]!=e[e.length-(n+1)])return t.slice(-n);return t.slice(-n)}function T(t,e,n){if(t.slice(0,e.length)!=e)throw Error(`string ${JSON.stringify(t)} doesn't start with prefix ${JSON.stringify(e)}; this is a bug`);return n+t.slice(e.length)}function k(t,e,n){if(!e)return t+n;if(t.slice(-e.length)!=e)throw Error(`string ${JSON.stringify(t)} doesn't end with suffix ${JSON.stringify(e)}; this is a bug`);return t.slice(0,-e.length)+n}function I(t,e){return T(t,e,"")}function S(t,e){return k(t,e,"")}function V(t,e){return e.slice(0,me(t,e))}function me(t,e){let n=0;t.length>e.length&&(n=t.length-e.length);let i=e.length;t.length<e.length&&(i=t.length);let r=Array(i),o=0;r[0]=0;for(let f=1;f<i;f++){for(e[f]==e[o]?r[f]=r[o]:r[f]=o;o>0&&e[f]!=e[o];)o=r[o];e[f]==e[o]&&o++}o=0;for(let f=n;f<t.length;f++){for(;o>0&&t[f]!=e[o];)o=r[o];t[f]==e[o]&&o++}return o}function W(t){let e;for(e=t.length-1;e>=0&&t[e].match(/\s/);e--);return t.substring(e+1)}function b(t){let e=t.match(/^\s*/);return e?e[0]:""}var N="a-zA-Z0-9_\\u{C0}-\\u{FF}\\u{D8}-\\u{F6}\\u{F8}-\\u{2C6}\\u{2C8}-\\u{2D7}\\u{2DE}-\\u{2FF}\\u{1E00}-\\u{1EFF}",de=new RegExp(`[${N}]+|\\s+|[^${N}]`,"ug"),z=class extends x{equals(e,n,i){return i.ignoreCase&&(e=e.toLowerCase(),n=n.toLowerCase()),e.trim()===n.trim()}tokenize(e,n={}){let i;if(n.intlSegmenter){let f=n.intlSegmenter;if(f.resolvedOptions().granularity!="word")throw new Error('The segmenter passed must have a granularity of "word"');i=Array.from(f.segment(e),s=>s.segment)}else i=e.match(de)||[];let r=[],o=null;return i.forEach(f=>{/\s/.test(f)?o==null?r.push(f):r.push(r.pop()+f):o!=null&&/\s/.test(o)?r[r.length-1]==o?r.push(r.pop()+f):r.push(o+f):r.push(f),o=f}),r}join(e){return e.map((n,i)=>i==0?n:n.replace(/^\s+/,"")).join("")}postProcess(e,n){if(!e||n.oneChangePerToken)return e;let i=null,r=null,o=null;return e.forEach(f=>{f.added?r=f:f.removed?o=f:((r||o)&&ee(i,o,r,f),i=f,r=null,o=null)}),(r||o)&&ee(i,o,r,null),e}},ne=new z;function B(t,e,n){return n?.ignoreWhitespace!=null&&!n.ignoreWhitespace?A(t,e,n):ne.diff(t,e,n)}function ee(t,e,n,i){if(e&&n){let r=b(e.value),o=W(e.value),f=b(n.value),s=W(n.value);if(t){let l=F(r,f);t.value=k(t.value,f,l),e.value=I(e.value,l),n.value=I(n.value,l)}if(i){let l=M(o,s);i.value=T(i.value,s,l),e.value=S(e.value,l),n.value=S(n.value,l)}}else if(n){if(t){let r=b(n.value);n.value=n.value.substring(r.length)}if(i){let r=b(i.value);i.value=i.value.substring(r.length)}}else if(t&&i){let r=b(i.value),o=b(e.value),f=W(e.value),s=F(r,o);e.value=I(e.value,s);let l=M(I(r,s),f);e.value=S(e.value,l),i.value=T(i.value,r,l),t.value=k(t.value,r,r.slice(0,r.length-l.length))}else if(i){let r=b(i.value),o=W(e.value),f=V(o,r);e.value=S(e.value,f)}else if(t){let r=W(t.value),o=b(e.value),f=V(r,o);e.value=I(e.value,f)}}var $=class extends x{tokenize(e){let n=new RegExp(`(\\r?\\n)|[${N}]+|[^\\S\\n\\r]+|[^${N}]`,"ug");return e.match(n)||[]}},te=new $;function A(t,e,n){return te.diff(t,e,n)}function ie(t,e){if(typeof t=="function")e.callback=t;else if(t)for(let n in t)Object.prototype.hasOwnProperty.call(t,n)&&(e[n]=t[n]);return e}var J=class extends x{constructor(){super(...arguments),this.tokenize=he}equals(e,n,i){return i.ignoreWhitespace?((!i.newlineIsToken||!e.includes(`
+`))&&(e=e.trim()),(!i.newlineIsToken||!n.includes(`
+`))&&(n=n.trim())):i.ignoreNewlineAtEof&&!i.newlineIsToken&&(e.endsWith(`
+`)&&(e=e.slice(0,-1)),n.endsWith(`
+`)&&(n=n.slice(0,-1))),super.equals(e,n,i)}},U=new J;function O(t,e,n){return U.diff(t,e,n)}function H(t,e,n){return n=ie(n,{ignoreWhitespace:!0}),U.diff(t,e,n)}function he(t,e){e.stripTrailingCr&&(t=t.replace(/\r\n/g,`
+`));let n=[],i=t.split(/(\n|\r\n)/);i[i.length-1]||i.pop();for(let r=0;r<i.length;r++){let o=i[r];r%2&&!e.newlineIsToken?n[n.length-1]+=o:n.push(o)}return n}function pe(t){return t=="."||t=="!"||t=="?"}var _=class extends x{tokenize(e){var n;let i=[],r=0;for(let o=0;o<e.length;o++){if(o==e.length-1){i.push(e.slice(r));break}if(pe(e[o])&&e[o+1].match(/\s/)){for(i.push(e.slice(r,o+1)),o=r=o+1;!((n=e[o+1])===null||n===void 0)&&n.match(/\s/);)o++;i.push(e.slice(r,o+1)),r=o+1}}return i}},re=new _;function q(t,e,n){return re.diff(t,e,n)}var G=class extends x{tokenize(e){return e.split(/([{}:;,]|\s+)/)}},oe=new G;function X(t,e,n){return oe.diff(t,e,n)}var Z=new Map([["diffChars",j],["diffWords",B],["diffWordsWithSpace",A],["diffLines",O],["diffTrimmedLines",H],["diffSentences",q],["diffCss",X]]);var se=(s=>(s.CHARS="diffChars",s.WORDS="diffWords",s.WORDS_WITH_SPACE="diffWordsWithSpace",s.LINES="diffLines",s.TRIMMED_LINES="diffTrimmedLines",s.SENTENCES="diffSentences",s.CSS="diffCss",s))(se||{}),fe=t=>{let e=t.split(`
+`);if(e.every(o=>!o))return e.length===2?[]:(e.pop(),e);let i=e[e.length-1],r=e[0];return i||e.pop(),r||e.shift(),e},xe=(t,e,n="diffChars")=>{let i=Object.values(se),r="diffChars";i.includes(n)&&Z.has(n)&&typeof Z.get(n)=="function"&&(r=n);let o=Z.get(r),f=o?o(t,e):[],s={left:[],right:[]};return f.forEach(({added:l,removed:c,value:u})=>{let a={};return l&&(a.type=1,a.value=u,s.right.push(a)),c&&(a.type=2,a.value=u,s.left.push(a)),!c&&!l&&(a.type=0,a.value=u,s.right.push(a),s.left.push(a)),a}),s},le=(t,e,n=!1,i="diffChars",r=0)=>{let o=O(t.trimRight(),e.trimRight(),{newlineIsToken:!0,ignoreWhitespace:!1}),f=r,s=r,l=[],c=0,u=[],a=[],D=(h,m,L,g,y,d)=>fe(h).flatMap((C,w)=>{let v={},p={};if(a.includes(`${m}-${w}`)||y&&w!==0)return[];if(L||g)if(u.includes(c)||u.push(c),g){s+=1,v.lineNumber=s,v.type=2,v.value=C||" ";let E=o[m+1];if(E&&E.added&&fe(E.value)[w]){let{value:Q,lineNumber:ae,type:ce}=D(E.value,m,!0,!1,!0,!0)[0].right;if(a.push(`${m+1}-${w}`),p.lineNumber=ae,p.type=ce,n)p.value=Q;else{let Y=xe(C,Q,i);p.value=Y.right,v.value=Y.left}}}else f+=1,p.lineNumber=f,p.type=1,p.value=C;else s+=1,f+=1,v.lineNumber=s,v.type=0,v.value=C,p.lineNumber=f,p.type=0,p.value=C;return c+=1,d||l.push({left:v,right:p,diffIndex:m+c-1}),[{right:p,left:v,diffIndex:m+c-1}]});return o.forEach(({added:h,removed:m,value:L},g)=>{D(L,g,h,m)}),{lineInformation:l,diffLines:u}};function ve({oldValue:t,newValue:e,disableWordDiff:n,compareMethod:i,linesOffset:r,extraLinesSurroundingDiff:o,showDiffOnly:f,expandedBlockIdsSet:s}){let{lineInformation:l,diffLines:c}=le(t,e,n,i,r),u=o<0?0:o,a=[],D=0,h=[];return l.forEach((m,L)=>{let g=c[D],y=g-L;if(f&&(y===-u&&(a=[],D+=1),m.left.type===0&&(y>u||typeof g>"u")&&!s.has(g))){if(a.push(L+1),L===l.length-1&&a.length>1){h.push({num:a.length,blockNumber:g,leftBlockLineNumber:m.left.lineNumber,rightBlockLineNumber:m.right.lineNumber,diffIndex:m.diffIndex});return}return null}if(y===u&&a.length>0){let{length:d}=a;a=[],h.push({num:d,blockNumber:g,leftBlockLineNumber:m.left.lineNumber,rightBlockLineNumber:m.right.lineNumber,diffIndex:m.diffIndex});return}h.push(m)}),h}var ue=ve;console.log("Loaded RDV Worker");self.onmessage=function(t){try{let e=t.data,n=ue(e);self.postMessage({success:!0,data:n})}catch(e){self.postMessage({success:!1,error:e.message})}finally{self.close()}};})();
