+++ src/getLinesToRenderWorker.ts
